backend/requirements.txt
websockets
python-dotenv
rapidfuzz
yt-dlp
fastapi
uvicorn
python-multipart
langchain-groq
tk
motor
pymongo
